requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.0
timm>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.8

# Utilities
tqdm>=4.65.0
omegaconf>=2.3.0
pyyaml>=6.0
wandb>=0.15.0

# Dataset Download
kaggle>=1.5.13
huggingface-hub>=0.16.0
datasets>=2.14.0

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0
